Synced up with develop branch jenkins-ignore (#1625)
* Synced up with develop b ranch jenkins-ignore

* Remove ItMiiConfigUpdateMulti.java and ItMiiConfigUpdateSecret.java jenkins-ignore
