+++ integration-tests/src/test/resources/model-in-image/ds_app/wlsdeploy/applications/META-INF/MANIFEST.MF
@@ -1,0 +1,3 @@
+Manifest-Version: 1.0
+Created-By: 1.8.0_201 (Oracle Corporation)
+
