+++ integration-tests/src/test/resources/model-in-image/ds_app/wlsdeploy/applications/META-INF/application.xml
@@ -1,0 +1,13 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!-- Copyright (c) 2020, Oracle Corporation and/or its affiliates.
+     Licensed under the Universal Permissive License v 1.0 as shown at https://oss.oracle.com/licenses/upl. -->
+<application xmlns="http://java.sun.com/xml/ns/j2ee" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" version="1.4">
+  <description>JSP 2.0 Expression Language Example</description>
+  <display-name>JSP 2.0 Expression Language Example</display-name>
+  <module>
+    <web>
+      <web-uri>ds_war</web-uri>
+      <context-root>ds_war</context-root>
+    </web>
+  </module>
+</application>
