+++ integration-tests/src/test/resources/model-in-image/ds_app/wlsdeploy/applications/META-INF/weblogic-application.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!-- Copyright (c) 2020, Oracle Corporation and/or its affiliates.
+     Licensed under the Universal Permissive License v 1.0 as shown at https://oss.oracle.com/licenses/upl. -->
+<weblogic-application xmlns="http://www.bea.com/ns/weblogic/90">
+</weblogic-application>
