+++ integration-tests/src/test/resources/model-in-image/ds_app/wlsdeploy/applications/ds_war/WEB-INF/web.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!-- Copyright (c) 2020, Oracle Corporation and/or its affiliates.
+     Licensed under the Universal Permissive License v 1.0 as shown at https://oss.oracle.com/licenses/upl. -->
+<web-app xmlns="http://java.sun.com/xml/ns/j2ee" version="2.4">
+</web-app>
