+++ integration-tests/src/test/resources/model-in-image/ds_app/wlsdeploy/applications/ds_war/WEB-INF/weblogic.xml
@@ -1,0 +1,13 @@
+<?xml version="1.0" encoding="ISO-8859-1"?>
+<!-- Copyright (c) 2020, Oracle Corporation and/or its affiliates.
+     Licensed under the Universal Permissive License v 1.0 as shown at https://oss.oracle.com/licenses/upl. -->
+<weblogic-web-app xmlns="http://www.bea.com/ns/weblogic/90">
+  <session-descriptor>
+     <timeout-secs>15</timeout-secs>
+     <invalidation-interval-secs>60</invalidation-interval-secs>
+  </session-descriptor>
+  <jsp-descriptor>
+    <page-check-seconds>1</page-check-seconds>
+    <verbose>true</verbose>
+  </jsp-descriptor>
+</weblogic-web-app>
